+++ src/redisenterprise/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+1.3.0
+- Added a new required property: PublicNetworkAccess for Cluster.
+- Updated the default value of AccessKeysAuthentication property for Database to 'Disabled'.
+
 1.2.3
 - Added breaking change warning for upcoming release